refactor: Product DTO의 패키지 변경
