+++ src/main/java/cart/service/ProductService.java
@@ -1,9 +1,9 @@
 package cart.service;
 
 import cart.domain.product.Product;
-import cart.dto.ProductCreateRequestDto;
-import cart.dto.ProductEditRequestDto;
-import cart.dto.ProductsResponseDto;
+import cart.dto.product.ProductCreateRequestDto;
+import cart.dto.product.ProductEditRequestDto;
+import cart.dto.product.ProductsResponseDto;
 import cart.exception.ProductNotFoundException;
 import cart.repository.product.ProductRepository;
 import org.springframework.stereotype.Service;
