+++ src/test/java/cart/controller/ProductIntegrationTest.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
 import cart.domain.product.Product;
-import cart.dto.ProductCreateRequestDto;
-import cart.dto.ProductEditRequestDto;
+import cart.dto.product.ProductCreateRequestDto;
+import cart.dto.product.ProductEditRequestDto;
 import cart.repository.product.ProductRepository;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
